reorganize file structure of testing files
